drop erosion and min width
