+++ src/places/region_growing_traversability_clustering.cpp
@@ -42,7 +42,6 @@
 #include <spark_dsg/edge_attributes.h>
 #include <spark_dsg/node_symbol.h>
 
-#include <cstdlib>
 #include <queue>
 
 #include "hydra/utils/timing_utilities.h"
@@ -73,10 +72,8 @@ void declare_config(RegionGrowingTraversabilityClustering::Config& config) {
   field(config.max_radius, "max_radius", "m");
   field(config.num_orientation_bins, "num_orientation_bins");
   field(config.use_diagonal_connectivity, "use_diagonal_connectivity");
-  field(config.min_connection_width_voxels, "min_connection_width_voxels");
   check(config.max_radius, GT, 0.0f, "max_radius");
   check(config.num_orientation_bins, GE, 3, "num_orientation_bins");
-  check(config.min_connection_width_voxels, GE, 1, "min_connection_width_voxels");
 }
 
 RegionGrowingTraversabilityClustering::RegionGrowingTraversabilityClustering(
@@ -134,21 +131,8 @@ VoxelSet RegionGrowingTraversabilityClustering::initializeVoxels(
   Eigen::Vector3f start_2d = start_position.cast<float>();
   start_2d.z() = 0.0f;
   const auto start_index = layer.globalIndexFromPoint(start_2d);
-
-  const size_t num_neighbors = config.use_diagonal_connectivity ? 8u : 4u;
-  const int erosion_radius = config.min_connection_width_voxels - 1;
-  if (erosion_radius <= 0) {
-    return growRegion(candidates, start_index, num_neighbors);
-  }
-
-  const VoxelSet core =
-      erodeCandidates(candidates, erosion_radius, config.use_diagonal_connectivity);
-  if (core.find(start_index) == core.end()) {
-    // Robot cell is not "wide" (e.g. hugging a wall); fall back to plain
-    // connectivity so we never drop all places for a frame.
-    return growRegion(candidates, start_index, num_neighbors);
-  }
-  return growConnectedWithMinWidth(candidates, core, start_index, num_neighbors);
+  const size_t num_neighbors = config.use_diagonal_connectivity ? 8 : 4;
+  return growRegion(candidates, start_index, num_neighbors);
 }
 
 VoxelMap RegionGrowingTraversabilityClustering::initializeRegions(
@@ -178,7 +162,7 @@ VoxelMap RegionGrowingTraversabilityClustering::initializeRegions(
 
 void RegionGrowingTraversabilityClustering::growRegions(VoxelSet& all_voxels,
                                                         VoxelMap& assigned_voxels) {
-  const size_t num_neighbors = config.use_diagonal_connectivity ? 8u : 4u;
+  const size_t num_neighbors = config.use_diagonal_connectivity ? 8 : 4;
 
   // Try to grow existing regions into the closest voxels to each region.
   const auto max_dist_sq = static_cast<float>(max_region_size_ * max_region_size_);
@@ -468,69 +452,6 @@ VoxelSet RegionGrowingTraversabilityClustering::growRegion(
     }
   }
 
-  return result;
-}
-
-VoxelSet RegionGrowingTraversabilityClustering::erodeCandidates(
-    const VoxelSet& candidates, int radius, bool use_diagonal) {
-  if (radius <= 0) {
-    return candidates;
-  }
-  VoxelSet eroded;
-  for (const auto& v : candidates) {
-    bool keep = true;
-    for (int dx = -radius; dx <= radius && keep; ++dx) {
-      for (int dy = -radius; dy <= radius; ++dy) {
-        if (dx == 0 && dy == 0) {
-          continue;
-        }
-        // Structuring element follows the connectivity: Chebyshev (square) ball for
-        // 8-connected, Manhattan (plus) ball for 4-connected.
-        const int reach = use_diagonal ? std::max(std::abs(dx), std::abs(dy))
-                                       : std::abs(dx) + std::abs(dy);
-        if (reach > radius) {
-          continue;
-        }
-        if (candidates.find(v + VoxelIndex(dx, dy, 0)) == candidates.end()) {
-          keep = false;
-          break;
-        }
-      }
-    }
-    if (keep) {
-      eroded.insert(v);
-    }
-  }
-  return eroded;
-}
-
-VoxelSet RegionGrowingTraversabilityClustering::growConnectedWithMinWidth(
-    const VoxelSet& candidates,
-    const VoxelSet& core,
-    const VoxelIndex& seed,
-    size_t num_neighbors) {
-  VoxelSet result;
-  if (candidates.find(seed) == candidates.end()) {
-    return result;
-  }
-  VoxelSet visited;
-  std::queue<VoxelIndex> queue;
-  queue.push(seed);
-  visited.insert(seed);
-  while (!queue.empty()) {
-    const auto current = queue.front();
-    queue.pop();
-    result.insert(current);
-    if (core.find(current) == core.end()) {
-      continue;  // narrow voxel: include but do not propagate through it
-    }
-    for (size_t k = 0; k < num_neighbors && k < neighbors_.size(); ++k) {
-      const VoxelIndex n = current + neighbors_[k];
-      if (candidates.find(n) != candidates.end() && visited.insert(n).second) {
-        queue.push(n);
-      }
-    }
-  }
   return result;
 }
